+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Move from Structured Data Testing Tool to Schema Markup Validator in Admin Panel (#66)
+
 ## [1.11.1] - 2021-04-06
 
 ### Fixed